Adds tests

+++ src/utils.cc
@@ -7,7 +7,7 @@ double utils::get_elapsed(std::chrono::time_point<std::chrono::steady_clock> sta
   return std::chrono::duration_cast<std::chrono::milliseconds>(end-start).count();
 }
 
-int utils::dist(Position pos1, Position pos2) {
+double utils::dist(Position pos1, Position pos2) {
   return std::sqrt(pow(pos2.first - pos1.first, 2) + pow(pos2.second - pos1.second, 2));
 }
