+++ src/utils.h
@@ -14,7 +14,7 @@ namespace utils {
   double get_elapsed(std::chrono::time_point<std::chrono::steady_clock> start,
     std::chrono::time_point<std::chrono::steady_clock> end);
 
-  int dist(Position pos1, Position pos2);
+  double dist(Position pos1, Position pos2);
 
   std::string PositionToString(Position pos);
